refactor: move ceil_log2 into steel_utils::mth
